Big otimization of the "status" operation in case of global "submit to source control" (17s to 1.5s)
In this case there is a great number of files from the Engine, reported as "not in a workspace".
Also Plastic fileinfo does not return any results when called with at least on file not in a workspace

Thus we group files by path (ie. by subdirectory), and then we stop running "status" commands at the first error on a batch.