Fix source control menu extender to work with UE5
The old FLevelEditorMenuExtender is not used anymore by the Editor, so calling it does nothing.

The official multi-user editing plugin "Concert" had to switch to the "ToolMenus" package,
so I borrowed the new method they switched to and adapted the source code.

edit: added the missing call to Unnregister(). I had to make it robust to double calls, happening when closing the Editor.

Also removed an old unused "PlasticSourceControlMenu" FName variable